chore(llma): rename ai-costs directory
