+++ plugin-server/src/ingestion/ai-events/providers/embeddings.ts
