+++ plugin-server/src/ingestion/ai-events/providers/generated-providers.json
