+++ plugin-server/src/ingestion/ai-events/providers/llm-provider-costs.json
