+++ plugin-server/src/ingestion/ai-events/providers/manual-providers.ts
