+++ plugin-server/src/ingestion/ai-events/scripts/update-ai-costs.ts
